Python target inactive logic

+++ src/nnUtils.py
@@ -4,6 +4,9 @@ import cv2
 import os
 import sys
 import numpy as np
+from multiprocessing import Process
+from threading import Thread
+
 from person import Person
 
 from keras.models import load_model, Model
@@ -19,11 +22,9 @@ session = tf.Session(config=config)
 
 EVENT_LOOP_DELAY = 40	# delay for GUI window
                         # 40 ms equates to 1000ms/25fps = 40ms per frame
-
 imagesFIFO = "/tmp/images.fifo"
-
 intsFIFO = "/tmp/ints.fifo"
-
+deletionFIFO = "/tmp/deletion.fifo"
 
 people = []
 
@@ -31,9 +32,6 @@ people = []
 
 def euclidean_distance(vects):
     x, y = vects
-
-    #return K.cast(K.less(K.sqrt(K.sum(K.square(x - y), axis=1, keepdims=True)),0.5),"float32")
-
     return K.sqrt(K.sum(K.square(x - y), axis=1, keepdims=True))
 
 
@@ -42,16 +40,13 @@ def eucl_dist_output_shape(shapes):
     return (shape1[0], 1)
 
 def contrastive_loss(y_true, y_pred):
-    '''Contrastive loss from Hadsell-et-al.'06
-    http://yann.lecun.com/exdb/publis/pdf/hadsell-chopra-lecun-06.pdf
-    '''
     margin = 1
     return K.mean(y_true * K.square(y_pred) + (1 - y_true) * K.square(K.maximum(margin - y_pred, 0)))
 
 def calc_accuracy(labels, predictions):
-    '''accuracy function for compilation'''
     return K.mean(K.equal(labels, K.cast(K.less(predictions,0.5),"float32")))
 
+#####################################################################
 
 def convertForKeras(img):
     newImg = np.array(img)
@@ -86,13 +81,14 @@ def whichPerson(img):
         closest = 100
         closestPerson = 100
         for person in people:
-            currentPerson = person.getIdentifier()
-            previous = person.getPrevious()
-            for previousFrame in previous:
-                prediction = queryNeuralNetwork(personROI,previousFrame)                
-                if prediction < closest:
-                    closest = prediction
-                    closestPerson = currentPerson
+            if not person.isActive():
+                currentPerson = person.getIdentifier()
+                previous = person.getPrevious()
+                for previousFrame in previous:
+                    prediction = queryNeuralNetwork(personROI,previousFrame)                
+                    if prediction < closest:
+                        closest = prediction
+                        closestPerson = currentPerson
         if closest < 0.5:
             person = people[closestPerson]
             person.addPrevious(personROI)
@@ -109,15 +105,35 @@ def write(person):
     fifo.close()
 
 def processImage():
-    print("FIFO opened")
     while True:
         fifo = open(imagesFIFO, "r")
         data = fifo.read()
-        img = cv2.imread(data)
-        person = whichPerson(img)
-        os.remove(data)
-        write(person)
+
+        temp = data.split("/")
+        if len(temp) > 1:
+            img = cv2.imread(data)
+            person = whichPerson(img)
+            write(person)
+            os.remove(data)
+        else:
+            idToDelete = int(data)
+            people[idToDelete].makeInactive()
         fifo.close()
+
+# def processDeletion():
+#     while True:
+#         fifo = open(deletionFIFO, "r")
+#         data = fifo.read()
+        
+#         idToDelete = int(data)
+
+#         print(len(people))
+
+#         print(idToDelete)
+
+#         people[idToDelete].makeInactive()
+
+#         fifo.close()
 
 
 model = load_model("saved_models/openWorld.h5", custom_objects={'contrastive_loss': contrastive_loss, 'calc_accuracy': calc_accuracy, 'euclidean_distance': euclidean_distance, 'eucl_dist_output_shape': eucl_dist_output_shape})
@@ -127,18 +143,27 @@ model = load_model("saved_models/openWorld.h5", custom_objects={'contrastive_los
 if __name__ == '__main__':
     try:
         os.mkfifo(imagesFIFO)
+    except:
+        print("image fifo exists")
+
+    try:
         os.mkfifo(intsFIFO)
     except:
-        print("fifos exist")
+        print("ints fifo exists")
 
+    try:
+        os.mkfifo(deletionFIFO)
+    except:
+        print("deletion fifo exists")
 
     try:
         os.mkdir("/tmp/imgs")
     except:
         print("directory exists")
 
+    # p1 = Thread(target=processImage)
+    # p1.start()
+    # p2 = Thread(target=processDeletion)
+    # p2.start()
+
     processImage()
-
-
-    os.remove(imagesFIFO)
-    os.remove(intsFIFO)
